Final Commit
Project is now rewritten in C# and the database migrated to MSSQL.

v3 is the last version and implemented an ML API to set the relays as required to adjust the environment.

This was simple project based on two fans and three DHT22.

The data trained for the model is included as climatesys.csv - It was trained using linear regression.

As part of the ML API, a matrix of all possible output was created to make the predictions of the given outcome for the system.